ENH: first draft of skeptical signal
This is a class for dealing with detectors which do not
have a trigger at the hardware level, but are being scanned
against anyway.  This adds a software trigger which requires that
at least one update from the IOC (which changes the timestamp) happen
before the detector reports it is done.