ENH: make _set_readback skeptical
This is to accont for IOCs which when encountering a bad
value chose to return 0 instead nan.

If the value coming out of a monitor subscription is 0.0, reject this
update and contiue with life.  This short-circuts all of the base-class
logic which updates the python-side value + timestamp and runs the
ophyd layer callbacks.