API: do not use global count time
As part of mocking the spec API a global count time
was added to be used in the case when a time is not passed in.  This
removes the use of that global time, instead detectors count for
their pre-configured time which can result in different detectors
counting for different times.